requirements.txt
django==4.2.10
djangorestframework==3.14.0
drf-spectacular==0.27.2
python-dotenv==1.0.1
psycopg[binary]==3.2.12
django-cors-headers==4.4.0
dj-database-url==2.1.0
requests==2.31.0

